+++ posthog-node/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.4.0 - 2024-01-09
+
+1. Numeric property handling for feature flags now does the expected: When passed in a number, we do a numeric comparison. When passed in a string, we do a string comparison. Previously, we always did a string comparison.
+2. Add support for relative date operators for local evaluation.
+
+# 3.3.0 - 2024-01-02
+
+1. Adds PostHogSentryIntegration to allow automatic capturing of exceptions reported via the @sentry/node package
+
 # 3.2.1 - 2023-12-15
 
 1. Fixes issue where a background refresh of feature flags could throw an unhandled error. It now emits to be detected by `.on('error', ...)`